+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 12.1.0 - 2021-09-16
+
 ### Improved
 - `process.env.NODE_ENV` is now set to `development` using a webpack `DefinePlugin`